move tests to src folder (#957)
