+++ src/test/test-properties.js
@@ -1,6 +1,6 @@
 require('./include/common');
-const JsSIP = require('../');
-const pkg = require('../package.json');
+const JsSIP = require('../..');
+const pkg = require('../../package.json');
 
 describe('Properties', () => {
 	test('should have a name property', () => {
